requirements.txt
asgiref==3.11.0
Django==6.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.4
tzdata==2025.2
gunicorn==23.0.0
whitenoise==6.11.0
Pillow==12.1.0



